Intial set of changes to enable autofill on Android.
This is the initial checkin of AutoFill code into external/chromium.
There's still lots to do, but this basically gets autofill working
end to end, using a precanned profile injected from WebCore. By
default, autofill is not enabled at the moment. For testing you can
turn it on by setting the ENABLE_AUTOFILL environment variable to
"true" and rebuilding.

There is a corresponding change in external/webkit.

Change-Id: I74ef1ccc3fd0625f1bb8e2f85b43a60d23c59ffc