Merge Chromium at 7.0.540.0 : Manual merge of OpenSSL code.
We had redundant work in the Android and Chrome trees, and git
failed to merge them. I'm intentionally replacing the Android
code with the Chrome code here. The Chrome code is still being
actively developed, and we want to be able to merge fixes from
there directly into Android.

x509_certificate_openssl.cc doesn't exist in the Chrome code
yet, but it will soon. Keeping it as a dummy implementation
for now.

Change-Id: I1204d6274ce312d76d0533288cd51182e3f16491