Add Android MIME types to Chrome HTTP stack.
This is needed for loading webarchive files, because Android needs
to have the MIME type set correctly. I've uncommented the calls to
PlatformMimeUtil, and implemented this class by calling through to
libcore.net.MimeUtils.

To test, save google.com as a web archive. This should be saved as
/sdcard/Download/index.webarchivexml. Then run this command:

adb shell am start \
    -d file:///sdcard/Download/index.webarchivexml \
    -n com.android.browser/.BrowserActivity \
    -a android.intent.action.VIEW

The browser should load the archived google.com.

Change-Id: I26b3700eb3ec793d8c0602feeb36aea2be9cd23e