Fix bug in URLFetcherProxy
The url() function only returns a valid URL once a response has been
received as it represents the final URL that was connected to after
any redirects. It's therefore not correct to use it in the DoStart
method of URLFetcherProxy. Instead, cache the url that gets passed
into the URLFetcherProxy and use that.

Change-Id: I509f5bc490362b45c5814fc7b1bc57168903b17b