Persist session cookies on Android.
The tricky bit here is that the SQLite schema assumes expires_utc
will always be set. Our SQL API doesn't have good support for null
columns, so I'm using a sentinel value of 0 for session cookies.

This seems reasonably clean, and avoids a schema change.

Most of the save/restore logic is in the Browser app. All we're
doing here is ensuring that session cookies *can* be saved.

To test that session cookies survive a clean browser exit:
1. Go to a site that sets a session cookie.
2. Wait 30 seconds for the cookie DB to be flushed.
3. Press the home button.
4. Kill the browser (adb shell kill <PID>)
5. Open the browser again; session should have survived.

To test that session cookies do *not* survive a forced exit:
1. Set a session cookie, wait 30 seconds.
2. Do NOT press the home button, just kill the browser.
3. Open the browser again; session should be destroyed.

It would be good to flush the cookie DB immediately when the
browser is paused, but I'll do that in a separate CL.

Bug: 2864791
Change-Id: I241b471b86b8b5bb6593c7e78f6f0b8c83850e48