Fix build. Switch to a new way of spoofing prctl.
The code so far relied on using a different definition
from the bionic one, so there was no conflict. However bionic
changed to move to the same definition, which caused a conflict
at the C++ level. This uses the preprocessor to eliminate the
conflict, by first including the headers that would be afected
by the preprocessor trick and then redefining the function in
question with the preprocessor.

None of the call sites rely on side effects of evaluating
parameters to prctl, so it's OK to have the preprocessor
swallow them.

Change-Id: Id33ec44d4faf0e50f4dcea4b3c073f6318e3bba5