Reduce pref usage by autofill
Try to avoid using prefs in Android. We are not using
the entire system, and we have significant local edits.

Will try to remove the local edits in prefs and
keep them in autofill. We should not loose any
functionality as there is no way for the user
to change the prefs.

Change-Id: I5ee38228fa552127a0399390a617e1a324e75d86