Merge Chromium at r12.0.742.93: Various profile.h compile fixes
Not including chrome/common/extensions/extension.h, the main
reason is that this includes UserScript, which crashes with
WebCore code. Forward declaring needed struct.

ifdef out GetResourceContext as it returns a const
content::ResourceContext&. We cannot create this
without including the file, which are not in the
android chromium tree.

Change-Id: I3540894d5e590ac7decea92c0ef0562e622222e5