Null pointer check missing in chromium
When clearing a struct a null pointer check is missing in one
location in chromium.
The null check is present in another similar use cases.

When the backend cache is reset a member of the struct data_
 is set to zero before data_ itself is set to null. If the
backend cache  previously has been reset data_ will already
be null and the member cannot be accessed
(or uses undefined memory). The added null check will prevent
the clearing of the member variable when data_ is already null.

Change-Id: I475fea7436c871c2e82b841a93874e13a05161f4